Added exiting process

+++ src/thread/exit_thread.c
@@ -2,4 +2,4 @@
 
 #include <los/syscall.h>
 
-void exit_thread(uint64_t status) { system_call1(EXIT_THREAD_SYSCALL, status); }
+void exit_thread(int64_t status) { system_call1(EXIT_THREAD_SYSCALL, (uint64_t)status); }
